Send session IdleHint to logind once screen saver blanks
If systemd or elogind support is enabled, set the session IdleHint
when the screen saver activates (blanks).

Note that this approach is not perfect, as it does not include
xscreensaver's authentication dialog. It is therefore possible that
the system enters the suspend mode while the user is typing its
password in xscreensaver's authentication dialog. However, the window
of opportunity is pretty small.

Thanks to Michael "mjt0k" Tokarev and damjan from #systemd for help
with setting the idle hint.